# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "berpar-Music",
        "contact": {}
    },
    "paths": {
...
        "/api/song/detail": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "根据歌曲id获得歌曲信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "歌曲id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/song/singer/detail": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "根据歌手id获得歌曲信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "歌手id",
                        "name": "singerId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/song/singerName/detail": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "根据歌手名称获得歌曲信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "歌手姓名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    }
}